+++ docker/pxs-schema.sql
@@ -12,7 +12,7 @@ create table if not exists asset_classes
   executor_id varchar not null,
   class_name varchar not null,
   description varchar,
-  foreign key (executor_id) references executors,
+  foreign key (executor_id) references executors on update cascade on delete cascade,
   primary key (executor_id, class_name)
 )
 ;
@@ -23,7 +23,7 @@ create table if not exists assets
   asset_id varchar not null,
   class_name varchar not null,
   description varchar,
-  foreign key (executor_id, class_name) references asset_classes,
+  foreign key (executor_id, class_name) references asset_classes on update cascade on delete cascade,
   primary key (executor_id, asset_id)
 )
 ;
@@ -33,7 +33,7 @@ create table if not exists mission_defs
   executor_id varchar not null,
   mission_def_id varchar not null,
   description varchar,
-  foreign key (executor_id) references executors,
+  foreign key (executor_id) references executors on update cascade on delete cascade,
   primary key (executor_id, mission_def_id)
 )
 ;
@@ -43,8 +43,8 @@ create table if not exists mission_def_asset_class
   executor_id varchar not null,
   mission_def_id varchar not null,
   asset_class_name varchar not null,
-  foreign key (executor_id, mission_def_id) references mission_defs,
-  foreign key (executor_id, asset_class_name) references asset_classes,
+  foreign key (executor_id, mission_def_id) references mission_defs on update cascade on delete cascade,
+  foreign key (executor_id, asset_class_name) references asset_classes on update cascade on delete cascade,
   primary key (executor_id, mission_def_id, asset_class_name)
 )
 ;
@@ -58,7 +58,7 @@ create table if not exists parameters
   required boolean default false not null,
   default_value varchar,
   description varchar,
-  foreign key (executor_id, mission_def_id) references mission_defs,
+  foreign key (executor_id, mission_def_id) references mission_defs on update cascade on delete cascade,
   primary key (executor_id, mission_def_id, param_name)
 )
 ;
@@ -73,8 +73,8 @@ create table if not exists missions
   start_date timestamp with time zone,
   end_date timestamp with time zone,
   geometry json,
-  foreign key (executor_id, mission_def_id) references mission_defs,
-  foreign key (executor_id, asset_id) references assets,
+  foreign key (executor_id, mission_def_id) references mission_defs on update cascade on delete cascade,
+  foreign key (executor_id, asset_id) references assets on update cascade on delete cascade,
   primary key (mission_id)
 )
 ;
@@ -87,7 +87,7 @@ create table if not exists arguments
   param_name varchar not null,
   param_value varchar not null,
   primary key (mission_id, param_name),
-  foreign key (mission_id) references missions (mission_id),
-  foreign key (executor_id, mission_def_id, param_name) references parameters
+  foreign key (mission_id) references missions (mission_id) on update cascade on delete cascade,
+  foreign key (executor_id, mission_def_id, param_name) references parameters on update cascade on delete cascade
 )
 ;